+++ src/lib/deskflow/KeyMap.h
@@ -9,7 +9,7 @@
 
 #include "base/String.h"
 #include "common/StdMap.h"
-#include "common/stdset.h"
+#include "common/StdSet.h"
 #include "common/stdvector.h"
 #include "deskflow/KeyTypes.h"
