+++ src/lib/net/SecureListenSocket.h
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "common/stdset.h"
+#include "common/StdSet.h"
 #include "net/SecurityLevel.h"
 #include "net/TCPListenSocket.h"
