+++ src/lib/platform/OSXKeyState.h
@@ -9,7 +9,7 @@
 
 #include "OSXAutoTypes.h"
 #include "common/StdMap.h"
-#include "common/stdset.h"
+#include "common/StdSet.h"
 #include "common/stdvector.h"
 #include "deskflow/KeyState.h"
