+++ src/lib/platform/XWindowsScreen.h
@@ -8,7 +8,7 @@
 #pragma once
 
 #include "arch/Arch.h"
-#include "common/stdset.h"
+#include "common/StdSet.h"
 #include "common/stdvector.h"
 #include "deskflow/ClientArgs.h"
 #include "deskflow/KeyMap.h"
