+++ src/lib/server/ClientListener.h
@@ -11,7 +11,7 @@
 #include "base/Event.h"
 #include "base/EventTypes.h"
 #include "common/StdDeque.h"
-#include "common/stdset.h"
+#include "common/StdSet.h"
 #include "net/SecurityLevel.h"
 #include "server/Config.h"
